Add permissions to the claim as well, restructure usecase and scope

+++ objectstore-server/src/http.rs
@@ -15,7 +15,7 @@ use sentry::integrations::tower as sentry_tower;
 use serde::Serialize;
 use uuid::Uuid;
 
-use crate::authentication::{Claim, ExtractScope};
+use crate::authentication::{Claim, ExtractScope, Permission};
 use crate::config::Config;
 use crate::state::ServiceState;
 
@@ -57,6 +57,7 @@ async fn put_blob_no_key(
     ExtractScope(claim): ExtractScope,
     body: Body,
 ) -> error::Result<impl IntoResponse> {
+    claim.ensure_permission(Permission::Write)?;
     let key = claim.into_key(Uuid::new_v4().to_string());
 
     let stream = body.into_data_stream().map_err(anyhow::Error::from).boxed();
@@ -72,6 +73,7 @@ async fn put_blob(
     Path(key): Path<String>,
     body: Body,
 ) -> error::Result<impl IntoResponse> {
+    claim.ensure_permission(Permission::Write)?;
     let key = claim.into_key(key);
 
     let stream = body.into_data_stream().map_err(anyhow::Error::from).boxed();
@@ -86,6 +88,7 @@ async fn get_blob(
     ExtractScope(claim): ExtractScope,
     Path(key): Path<String>,
 ) -> error::Result<Response> {
+    claim.ensure_permission(Permission::Read)?;
     let key = claim.into_key(key);
 
     let Some(contents) = state.service.get_file(&key).await? else {
@@ -101,6 +104,7 @@ async fn delete_blob(
     ExtractScope(claim): ExtractScope,
     Path(key): Path<String>,
 ) -> error::Result<impl IntoResponse> {
+    claim.ensure_permission(Permission::Write)?;
     let key = claim.into_key(key);
 
     state.service.delete_file(&key).await?;
@@ -114,23 +118,31 @@ mod error {
     use axum::http::StatusCode;
     use axum::response::{IntoResponse, Response};
 
-    pub struct AnyhowError(anyhow::Error);
+    pub enum AnyhowResponse {
+        Error(anyhow::Error),
+        Response(Response),
+    }
 
-    pub type Result<T> = std::result::Result<T, AnyhowError>;
+    pub type Result<T> = std::result::Result<T, AnyhowResponse>;
 
-    impl IntoResponse for AnyhowError {
+    impl IntoResponse for AnyhowResponse {
         fn into_response(self) -> Response {
-            eprintln!("{:?}", self.0); // TODO: this is just temporary
-            StatusCode::INTERNAL_SERVER_ERROR.into_response()
+            match self {
+                AnyhowResponse::Error(_error) => StatusCode::INTERNAL_SERVER_ERROR.into_response(),
+                AnyhowResponse::Response(response) => response,
+            }
         }
     }
 
-    impl<E> From<E> for AnyhowError
-    where
-        E: Into<anyhow::Error>,
-    {
-        fn from(err: E) -> Self {
-            Self(err.into())
+    impl From<Response> for AnyhowResponse {
+        fn from(response: Response) -> Self {
+            Self::Response(response)
+        }
+    }
+
+    impl From<anyhow::Error> for AnyhowResponse {
+        fn from(err: anyhow::Error) -> Self {
+            Self::Error(err)
         }
     }
 }
